requirements.txt
python-dotenv
streamlit
pandas
openai
tabulate